+++ ui/appui-react/src/appui-react/statusfields/SnapMode.tsx
@@ -23,6 +23,8 @@ import { SvgSnapsMidpoint } from "../icons/snaps/SvgSnapsMidpoint.js";
 import { SvgSnapsNearest } from "../icons/snaps/SvgSnapsNearest.js";
 import { SvgSnapsOrigin } from "../icons/snaps/SvgSnapsOrigin.js";
 import { SvgSnaps } from "../icons/snaps/SvgSnaps.js";
+import { SvgSnapsPerpendicular } from "../icons/snaps/SvgSnapsPerpendicular.js";
+import { SvgSnapsTangent } from "../icons/snaps/SvgSnapsTangent.js";
 
 /** Define the properties that will be used to represent the available snap modes. */
 interface SnapModeFieldEntry {
@@ -68,6 +70,28 @@ const allSnapModeFieldEntries: SnapModeFieldEntry[] = [
     value: SnapMode.Bisector as number,
     icon: <SvgSnapsBisector />,
   },
+  // @ts-ignore Introduced in @itwin/core-frontend@5.0.0
+  ...(SnapMode.PerpendicularPoint === undefined
+    ? []
+    : [
+        {
+          labelKey: "snapModeField.perpendicularPoint",
+          // @ts-ignore Introduced in @itwin/core-frontend@5.0.0
+          value: SnapMode.PerpendicularPoint as number,
+          icon: <SvgSnapsPerpendicular />,
+        },
+      ]),
+  // @ts-ignore Introduced in @itwin/core-frontend@5.0.0
+  ...(SnapMode.TangentPoint === undefined
+    ? []
+    : [
+        {
+          labelKey: "snapModeField.tangentPoint",
+          // @ts-ignore Introduced in @itwin/core-frontend@5.0.0
+          value: SnapMode.TangentPoint as number,
+          icon: <SvgSnapsTangent />,
+        },
+      ]),
 ];
 
 /** Defines properties supported by the SnapMode Field Component. */