+++ ui/appui-react/src/appui-react/toolbar/new-toolbars/Item.tsx
@@ -45,6 +45,7 @@ export const Item = React.forwardRef<HTMLButtonElement, ItemProps>(
         labelProps={labelProps}
         style={props.style}
         ref={ref}
+        data-item-id={item.id}
         {...other}
       >
         {item.iconNode ?? (